Refactor SafeMath16 library to use uint16 for arithmetic operations

+++ contracts/SafeMath16.sol
@@ -10,11 +10,11 @@ library SafeMath16 {
   /**
   * @dev Multiplies two numbers, throws on overflow.
   */
-  function mul(uint32 a, uint32 b) internal pure returns (uint32) {
+  function mul(uint16 a, uint16 b) internal pure returns (uint16) {
     if (a == 0) {
       return 0;
     }
-    uint32 c = a * b;
+    uint16 c = a * b;
     assert(c / a == b);
     return c;
   }
@@ -22,9 +22,9 @@ library SafeMath16 {
   /**
   * @dev Integer division of two numbers, truncating the quotient.
   */
-  function div(uint32 a, uint32 b) internal pure returns (uint32) {
+  function div(uint16 a, uint16 b) internal pure returns (uint16) {
     // assert(b > 0); // Solidity automatically throws when dividing by 0
-    uint32 c = a / b;
+    uint16 c = a / b;
     // assert(a == b * c + a % b); // There is no case in which this doesn't hold
     return c;
   }
@@ -32,7 +32,7 @@ library SafeMath16 {
   /**
   * @dev Subtracts two numbers, throws on overflow (i.e. if subtrahend is greater than minuend).
   */
-  function sub(uint32 a, uint32 b) internal pure returns (uint32) {
+  function sub(uint16 a, uint16 b) internal pure returns (uint16) {
     assert(b <= a);
     return a - b;
   }
@@ -40,8 +40,8 @@ library SafeMath16 {
   /**
   * @dev Adds two numbers, throws on overflow.
   */
-  function add(uint32 a, uint32 b) internal pure returns (uint32) {
-    uint32 c = a + b;
+  function add(uint16 a, uint16 b) internal pure returns (uint16) {
+    uint16 c = a + b;
     assert(c >= a);
     return c;
   }
